Move Happo from Tools & Frameworks to Online services
While old Happo continues to be a thing, new development is done on the
happo.io service. Here are some of the things that are improved in new
Happo:

  - New Happo has support for multiple browsers.
  - Faster screenshooting with parallell rendering.
  - An enhanced diff review UI
  - Better integration with github.com

To prevent driving traffic to the old project which is only lightly
maintained these days, I'm updating the link. I'm also moving the entry
to the "Online services" section which is a better fit for happo.io.